feat: move books state back to App comp, allow Book comp customization
Trying to keep Book component concerns separated from Bookshelf. This is the first try at it, but for now, I have opened Book component for customization by rendering its children as part of the top content, after the thumbnail. This is in turn used by the BooksGrid which receives a bookTopContent function prop that is called while rendering each Book to set its children content. BooksGrid is in turn used by Bookshelf and SearchBooks components, so they can each set bookTopContent to render whatever they want.

I'm not sure I will keep this approach, as the customization it is allowing is really not needed in this project at this point (Bookshelf and SearchBooks component are rendering the same BookshelfChanger component, just changing how the shelf of a given book is found.